psbt: move input into new file
